typed obdd, adf still using usizes
