+++ src/lib.rs
@@ -34,4 +34,4 @@
 pub mod adf;
 pub mod datatypes;
 pub mod obdd;
-pub mod obdd2;
+//pub mod obdd2;
